Ex6 - Allow to tokenize all the rules for one term
I have a lot of difficulty getting the tokenizer to form the proper groupings.
I want to be able to do a simple cartesian product at the end to compile the grammar.
For this to work, each rule as to be well defined it it's grouping so alternative options
don't combine with each other
(i.e. '<rule> ::= 0 | 1 | 2 | 3 | 4' expands to '(1 2 3 4) and not '(1234))
single-tokens and flatten-tokens are remnants of those grouping difficulties.
I will try to refactor the groupings properly so they are well defined and easier to follow